refactor(rust-api): extract events/SSE subsystem into events.rs (sc-1636)
Gather the scattered event-stream pieces into events.rs: the EventMessage/
EventHub/EventHubState/EventTicketStore/EventTicket types and their impls,
the SSE handlers (create_event_ticket, job_events, sse_event_stream,
ready/message/heartbeat events), and prune_tickets.

The publish/publish_queue helpers and the EVENT_* consts stay at the
root (publish is called from every route; events.rs sees the consts via
`use super::*`). pub(crate) on the items reached from the root —
EventHub/EventTicketStore (AppState fields + create_app), EventMessage
(root publish constructs it), the publish/subscribe/new methods, and the
create_event_ticket/job_events handlers. tests.rs's EventHub/EventMessage
imports are repointed to super::events. Behavior-preserving.
lib.rs 6,006 -> 5,865 lines.
